Wrap RequestDealer up
